IN-337: refactor: logic for adding filtering has been moved to hook
